Check devices in connection test
Summary:
Previously this was checking that a client can successfully connect.
But that's useless for a user if flipper doesn't also detect the device running the app, or nothing will be shown on screen.

This checks that the device is also known about.

I've removed the `done.fail()` because if this gets called, you lose any failed expect messages.

Reviewed By: passy

Differential Revision: D13194856

fbshipit-source-id: 5461dc1bedb2f9af80ad0dad6780fc25a428db17