Throw error if more than one device has same csr in container
Summary:
Getting more visibility on this.

We're seeing cases where the wrong deviceId is getting stored in the device connection_config.json.

This comes from flipper desktop in the response of 'signCertificate', and is figured out using getTargetDeviceId.

If we ever see these logs then something is wrong, and seeing the csr should help.

Reviewed By: passy

Differential Revision: D13357882

fbshipit-source-id: eb6befe07e7f9e7606e8940a1fed619351946d37