Fire-and-forget for openFile
Summary:
`xdg-open` is blocking on Linux. That means we won't do our cleanup steps
(or reset the button for that matter) until the user closes their viewer app.

This changes the call to be a `spawn`, so we can continue the chain. As no
one had actually used the value captured in the promise, this shouldn't
make a difference API-wise.

Reviewed By: danielbuechele

Differential Revision: D13465571

fbshipit-source-id: 5f698916157906d76ab983109deb9abe142baa9e