Explicitly start adb server
Summary:
We've only got about a 75% success rate at initializing an adb client: https://fburl.com/scuba/tr5tkehi

This is the second attempt to fix this. The first one D13487864 was reverted because it used 'adb' instead of the full path, and when run as a GUI app, you don't get the PATH variable typically set up in .bashrc files etc.

Fixing it by using the full $ANDROID_HOME path.

For facebook users, this will always be guessed correctly, but for OSS it could be installed anywhere, so we might get it wrong - it's common for $ANDROID_HOME to not be set at all on mac systems, and setting it for GUI apps is not exactly trivial: https://superuser.com/questions/476752/setting-environment-variables-in-os-x-for-gui-applications

Leaving the previous implementation as a fallback for OSS users and in case this method fails for some reason.

Metrics wrapped around each implementation as well as the entire thing, so we can see if it solves the issue.

Reviewed By: passy

Differential Revision: D13818981

fbshipit-source-id: d90bfd9ec80621526accc080778afc44f2a099d9